Major Update.
Moved all options for SoundManager.getInstance() and for playSound() to be options objects, instead of having lots and lots of arguments, that have to be in the right order.

Got rid of interrupt system. Now you assign each sound a priority number, and that decides whether a new sound should interrupt an old one or not.

Moved the classpath to com.primalscreen.utils.SoundManager

Made the second argument in playSound() a reference to 'this,' the calling class, so that the new function, stopSoundsFrom(this), could match all sounds from the same class and kill them (good for cleaning up before destroying things.

Made the docs say "argument" where they should. (before I was saying "property", oops).